feat: move scan mem/file ranges to util and reuse
