+++ src/mito2/src/read.rs
@@ -22,6 +22,7 @@ pub mod projection;
 pub(crate) mod prune;
 pub(crate) mod range;
 pub(crate) mod scan_region;
+pub(crate) mod scan_util;
 pub(crate) mod seq_scan;
 pub(crate) mod unordered_scan;
