+++ src/mito2/src/read/scan_util.rs
@@ -1,0 +1,136 @@
+// Copyright 2023 Greptime Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Utilities for scanners.
+
+use std::sync::{Arc, Mutex};
+use std::time::{Duration, Instant};
+
+use async_stream::try_stream;
+use futures::Stream;
+
+use crate::error::Result;
+use crate::read::range::RowGroupIndex;
+use crate::read::scan_region::StreamContext;
+use crate::read::{Batch, ScannerMetrics, Source};
+use crate::sst::parquet::reader::ReaderMetrics;
+
+struct PartitionMetricsInner {
+    metrics: ScannerMetrics,
+    reader_metrics: ReaderMetrics,
+}
+
+impl Drop for PartitionMetricsInner {
+    fn drop(&mut self) {
+        self.metrics.observe_metrics();
+    }
+}
+
+/// Metrics while reading a partition.
+#[derive(Clone)]
+pub(crate) struct PartitionMetrics(Arc<Mutex<PartitionMetricsInner>>);
+
+impl PartitionMetrics {
+    pub(crate) fn from_metrics(metrics: ScannerMetrics) -> Self {
+        let inner = PartitionMetricsInner {
+            metrics,
+            reader_metrics: ReaderMetrics::default(),
+        };
+        Self(Arc::new(Mutex::new(inner)))
+    }
+
+    pub(crate) fn inc_num_mem_ranges(&self, num: usize) {
+        let mut inner = self.0.lock().unwrap();
+        inner.metrics.num_mem_ranges += num;
+    }
+
+    pub(crate) fn inc_num_file_ranges(&self, num: usize) {
+        let mut inner = self.0.lock().unwrap();
+        inner.metrics.num_file_ranges += num;
+    }
+
+    pub(crate) fn inc_build_reader_cost(&self, cost: Duration) {
+        let mut inner = self.0.lock().unwrap();
+        inner.metrics.build_reader_cost += cost;
+    }
+
+    pub(crate) fn merge_metrics(&self, metrics: &ScannerMetrics) {
+        let mut inner = self.0.lock().unwrap();
+        inner.metrics.merge_from(metrics);
+    }
+
+    pub(crate) fn merge_reader_metrics(&self, metrics: &ReaderMetrics) {
+        let mut inner = self.0.lock().unwrap();
+        inner.reader_metrics.merge_from(metrics);
+    }
+}
+
+/// Scans memtable ranges at `index`.
+pub(crate) fn scan_mem_ranges(
+    stream_ctx: Arc<StreamContext>,
+    part_metrics: PartitionMetrics,
+    index: RowGroupIndex,
+) -> impl Stream<Item = Result<Batch>> {
+    try_stream! {
+        let ranges = stream_ctx.build_mem_ranges(index);
+        part_metrics.inc_num_mem_ranges(ranges.len());
+        for range in ranges {
+            let build_reader_start = Instant::now();
+            let iter = range.build_iter()?;
+            part_metrics.inc_build_reader_cost(build_reader_start.elapsed());
+
+            let mut source = Source::Iter(iter);
+            while let Some(batch) = source.next_batch().await? {
+                yield batch;
+            }
+        }
+    }
+}
+
+/// Scans file ranges at `index`.
+pub(crate) fn scan_file_ranges(
+    stream_ctx: Arc<StreamContext>,
+    part_metrics: PartitionMetrics,
+    index: RowGroupIndex,
+    read_type: &'static str,
+) -> impl Stream<Item = Result<Batch>> {
+    try_stream! {
+        let mut reader_metrics = ReaderMetrics::default();
+        let ranges = stream_ctx
+            .build_file_ranges(index, &mut reader_metrics)
+            .await?;
+        part_metrics.inc_num_file_ranges(ranges.len());
+        for range in ranges {
+            let build_reader_start = Instant::now();
+            let reader = range.reader(None).await?;
+            part_metrics.inc_build_reader_cost(build_reader_start.elapsed());
+            let compat_batch = range.compat_batch();
+            let mut source = Source::PruneReader(reader);
+            while let Some(mut batch) = source.next_batch().await? {
+                if let Some(compact_batch) = compat_batch {
+                    batch = compact_batch.compat_batch(batch)?;
+                }
+                yield batch;
+            }
+            if let Source::PruneReader(mut reader) = source {
+                reader_metrics.merge_from(reader.metrics());
+            }
+        }
+
+        // TODO(yingwen): Should we report metrics here?
+        // Reports metrics.
+        reader_metrics.observe_rows(read_type);
+        part_metrics.merge_reader_metrics(&reader_metrics);
+    }
+}
